+++ domain/review.ts
@@ -1,5 +1,14 @@
 import { State as FsrsState } from 'ts-fsrs';
+import { formatLocalDate } from './calendar';
 import type { Card } from './cards';
+
+export function isDueByDate(card: Card, referenceDate: Date, dayStartHour: number = 0): boolean {
+  const dueDate = new Date(card.fsrs.due);
+
+  const referenceDateStr = formatLocalDate(referenceDate, dayStartHour);
+  const dueStr = formatLocalDate(dueDate, dayStartHour);
+  return dueStr <= referenceDateStr;
+}
 
 const sortByDueDateThenSlug = (a: Card, b: Card): number => {
   const dueDiff = a.fsrs.due.getTime() - b.fsrs.due.getTime();
@@ -30,4 +39,11 @@ export function buildReviewQueue(
   allQueueCards.sort(sortByDueDateThenSlug);
 
   return allQueueCards;
+}
+
+export function calculateDelayedDueDate(due: Date, days: number): Date {
+  const currentDueDate = new Date(due);
+  const newDueDate = new Date(currentDueDate);
+  newDueDate.setDate(newDueDate.getDate() + days);
+  return newDueDate;
 }
